chore: pin old projen version (#18)
i'm investigating why the auto-approve workflow is failing, and one
possible explanation is that we are using an incompatible version of
`hmarr/auto-approve-action` in the latest projen version. pinning to an
older version that uses `2.2.1` that I can at least see has been
successful in other cdklabs repos. the aim is to revert back to the
latest version when I figure out the root cause.

all code changes comes from pinning the `projen` version and running
`npx projen`